Create systemd service to rescan the PCIe bus
At least on RHEL7, when the FPGA is loaded a further reboot or a PCI bus
rescan is needed to get the PCI registers properly mapped.  This service
is designed to be run during startup to rescan the bus.

At present the kernel driver has been observed entering a non-functional
state after using this service, so for the time being it is high risk.