Two changes to ensure driver upgrade actually upgrades!
First of all, disable weak module support from dkms.  Otherwise a "weak
module" is left behind pointing to the old version when upgrading this
driver, and this may be used instead of the updated driver.

Secondly, ask dracut (used to build the initramfs on RHEL) to not include
the driver in the initramfs.  It doesn't need to be loaded this early, and
again this means that the old version of the driver will be used until the
kernel is updated.